src/AspireRagDemo.UI/requirements.txt
streamlit>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.0
opentelemetry-distro[otlp]
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-http>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-api>=1.21.0
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-distro
opentelemetry-exporter-otlp